backend/requirements/base.txt
Django==2.2.3
django-environ==0.4.4
social-auth-app-django==2.1.0
django-model-utils==3.2.0
pytz==2019.1
django-timezone-field==3.0
stripe==2.32.1
django-fsm==2.6.1
psycopg2-binary==2.8.3
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
kombu==4.6.3
redis==3.2.1
vine==1.3.0
certifi==2019.6.16
chardet==3.0.4
idna==2.8
requests==2.22.0
urllib3==1.25.3
strawberry-graphql==0.13.3
